+++ examples/routes-config/src/pages/sales/favorites.tsx
@@ -1,0 +1,7 @@
+export default function Favorites() {
+  return (
+    <>
+      my favorite items
+    </>
+  );
+}