+++ packages/ice/src/getWatchEvents.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 import type { Context } from 'build-scripts';
-import type { Config } from '@ice/webpack-config/esm/types';
+import type { Config } from '@ice/webpack-config/types';
 import type { WatchEvent } from './types/plugin.js';
-import { generateRoutesInfo } from './routes.js';
+import { generateRoutesInfo, getRoutesDefination } from './routes.js';
 import type Generator from './service/runtimeGenerator';
 import getGlobalStyleGlobPattern from './utils/getGlobalStyleGlobPattern.js';
 import renderExportsTemplate from './utils/renderExportsTemplate.js';
@@ -19,16 +19,18 @@ interface Options {
   cache: Map<string, string>;
   ctx: Context<Config>;
   routeManifest: RouteManifest;
+  lazyRoutes: boolean;
 }
 
 const getWatchEvents = (options: Options): WatchEvent[] => {
-  const { generator, targetDir, templateDir, cache, ctx, routeManifest } = options;
+  const { generator, targetDir, templateDir, cache, ctx, routeManifest, lazyRoutes } = options;
   const { userConfig: { routes: routesConfig, dataLoader }, configFile, rootDir } = ctx;
   const watchRoutes: WatchEvent = [
     /src\/pages\/?[\w*-:.$]+$/,
     async (eventName: string) => {
       if (eventName === 'add' || eventName === 'unlink' || eventName === 'change') {
         const routesRenderData = await generateRoutesInfo(rootDir, routesConfig);
+        const { routeImports, routeDefination } = getRoutesDefination(routesRenderData.routes, lazyRoutes);
         const stringifiedData = JSON.stringify(routesRenderData);
         if (cache.get('routes') !== stringifiedData) {
           cache.set('routes', stringifiedData);
@@ -38,7 +40,7 @@ const getWatchEvents = (options: Options): WatchEvent[] => {
             generator.renderFile(
               path.join(templateDir, 'routes.ts.ejs'),
               path.join(rootDir, targetDir, 'routes.ts'),
-              routesRenderData,
+              { routeImports, routeDefination },
             );
             // Keep generate route manifest for avoid breaking change.
             generator.renderFile(