+++ packages/ice/tests/fixtures/nodeRunner/basic/entry.js
@@ -1,0 +1,3 @@
+const path = await __ice_import__('path');
+const a = await __ice_import__('./a.js');
+__ice_exports__.default = a.default;