+++ packages/plugin-icestark/runtime.d.ts
@@ -1,0 +1,2 @@
+export * from './esm/runtime/child';
+export * from './esm/runtime/framework';