+++ packages/route-manifest/src/routes.ts
@@ -151,7 +151,7 @@ function stripFileExtension(file: string) {
   return file.replace(/\.[a-z0-9]+$/i, '');
 }
 
-function createComponentName(id: string) {
+export function createComponentName(id: string) {
   return id.replace('.', '/') // 'pages/home.news' -> pages/home/news
     .split('/')
     .map((item: string) => item.toLowerCase())