+++ tests/integration/basic-project.test.ts
@@ -149,6 +149,7 @@ describe(`start ${example}`, () => {
 
   test('ClientOnly Component', async () => {
     await page.push('/client-only');
+    await page.waitForNetworkIdle();
     expect(await page.$$text('#mounted')).toStrictEqual(['Client']);
     const pageUrlText = await page.$$text('#page-url');
     expect((pageUrlText as string[])[0].endsWith('/client-only')).toBeTruthy();